\name{LncPathEnvir}
\alias{LncPathEnvir}
\docType{data}

\title{The variables in the environment variable \code{LncPathEnvir} of the system}
\description{
  The variables in the environment variable \code{LncPathEnvir} of the system.
}

\format{An environment variable}
\author{Junwei Han <hanjunwei1981@163.com>, Zeguo Sun <zeguo.sun@163.com>} 
\keyword{file}
